Fixed mrgdv issue with large number of files
mrgdv used to open all POTx files simulatiously to read headers
and data in two seperate loops. However, some kernels only allow
1024 files to be open at the same time, freezing mrgdv if the
number of files is larger.

Merging the two loops, opening on file at a time and closing it
right after header and data are read and written fixes this issue.
The only drawback is that a corrupt file at the end of the loop
could waste a lot of time.