Implemented test for the new 'prtlincompmatrixelements' feature of optic
- The test implemented (v9/t100) is exactly the same as v9/t14 but
with the aforementioned flag set to 1. It tests that the other features
of optic still works even though the flag is set to 1.
- It does not check that the matrix elements are well printed however
since that would require to test the _OPTIC.nc file. Though it is
possible, using simple python scripts, to check that the printed
elements can be used to reproduce exaclty what has been computed by the
optic utility.