Remove optional (front) argument in copy_on_gpu and copy_to_gpu.
When using an optional argument as the first argument, compiler generates a
warning, and it is also kind of misleading (looking for troubles).

Here it doesn't hurt to make the argument non-optional; it is actually required
for the GPU version, and it is better to have the same interface as possible
between cpu and gpu implementation of the same routine.