Add fortran function gpu_allocated.
This is a directly equivalent for fortran "allocated" but checking a gpu
pointer to check if it actually points to device allocated memory.

Strangely, I can't manage to bind as a fortran function (via iso_c_binding) a
cuda function returning a bool value (logical of kind c_bool). It should work
but the compiler is complaining is a way I can't fix. So that's why, here we
map the cuda function as a subroutine, and then make it a function on the
fortran side. A bit verbous for now....


Former-commit-id: fb65f477e4043b69e290376e8d209aa65cc69912