Fix a double-free error.
We can't have a call to unsetdevice_cuda when kokkos is enabled; kokkos is
responsible for freeing resources.

At some point, there should a clarifiication of the way cuda is initialized and
device selected in abinit. It should be much simpler; on a supercomputer the
batche scheduler set variables to control MPI/GPU affinity; on a node with
multiple GPU, the job scheduler sets CUDA_VISIBLE_DEVICES env variable to tell
the process which and how many gpus are available.


Former-commit-id: 991ec688ba0e2d2d201784cb05ee051dee09a496