Got rid of the inner loop over two perturbations but kept the infrastructure. Now, we harvest the occupancies pertaining to the unperturbed system, whose wavefunctions are read in. The unscreened occupations are stored in occ1 and the screened are stored in occ2 (even though, for the unperturbed case, they are the same). The perturbed case proceeds as normal, filling (unscreened) occ3 and (screened) occ4. By the time pawuj_det is called, nothing has changed.
Just some debugging print statements in 44_abitypes_defs/m_dtset.F90. Minor debugging in 98_main/ujdet.F90.

In 94_scfcv/m_scfcv_core.F90, pawuj_red is called twice now; once before the perturbation is applied and once after.

In 95_drive/m_gstate.F90, I've modified pawuj_drive to get rid of the inner loop over two perturbations and added diemixmag to the dtpawuj type because it will be needed for the calculation of the Hund's J parameter in macro_uj=4, which has officially been debugged and is ready to go.

The bulk of the modifications are in 65_paw/m_paw_uj.F90. I added a do loop in pawuj_red that does two iterations for the unperturbed calculation (repeats all settings for iuj=1 and iuj=2). If it's the second time pawuj_red has been called, then it proceeds as normal. There will be a difference in the output files for the test case, though, because vsh2 and occ2 will be printed much higher up in the file, and they will have the same occupations and perturbations as vsh1 and occ1. We will need to modify the test case accordingly.


Former-commit-id: ff4683ad76ffbeb00d2e62285e843d91862ecbf4