TMP/BROKEN Tweak ABINIT buildsystem to use a dedicated CMake for 44_manage_kokkos
Basically, CMake is a prerequisite of Kokkos as it is hardly feasible to
handle its complicate toolchain without it.

To adress this issue, we move all Kokkos C++ files and Fortran interface
modules in one subdirectory/library of ABINIT, and tweak generated top
Makefile.am so CMake is used instead of regular Makefile generated by
makemake and configure.

Doing this triggers the CMake step at "make" step, which is
counter-intuitive for a regular user... It is a first step and
at least it works. Also, no bother occurs if Kokkos isn't used at all.

TODO:
- CMake buildsystem is broken due to hard tweak on include paths for now
- "make install" target is broken as well, further tweak is required
- Need to check building outside of source directory

Former-commit-id: ee2790676cc8361e25948dde894129ab7e93e1c0