Add new input parameter "distribute_gemm_nonlop", for devs
This new parameter "force" the use of the distributed mode of GEMM
nonlop, if it is ever used.
It's mostly aimed for development and testing purposes as only really
huge case requires the distribution to save memory.

Right now, we don't study memory constraints to enable distribute
automatically but it should go that way.


Former-commit-id: 9c6c4276662b9e20420d9c00b740b3fb042dffaf