Fix some memory issues around invovl and vectproj
Running the testfarm highlighted a case when running some GCC version
would cause a failure because cprjin_ isn't assosciated in m_nonlop when
calling gemm_nonlop.

Another issue, occuring with some GCC version would cause seg fault
because of the use of optional vectproj, still when calling gemm_nonlop.
This is fixed by qualifying vectproj as ABI_CONTIGUOUS, as it is in
gemm_nonlop.


Former-commit-id: 9ced89550cd5596f2ffdf3d6a57ccebfa09dd653