WIP Testing some tests for OpenMP GPU (8 variants of Au31 + 2Si + 2Ni2O)
Disabling the rest for now.
The 8 variants of Au31 are for :
- istwfk    = [1,2]
- wfoptalg  = [111,114]
- paral_kgb = [0,1]

When paral_kgb=1, test are run on 1 and 2 MPI tasks (and GPUs).

Au31 comes from hpc folder and is what I'm used to but it's too heavy for light smoke
testing so it might belong to a separate testsuite once things are
settled.


Former-commit-id: 369361fd86e7cc7064de398cf4af674b197074dc