Debug phonons and el-ph
Explanation of the bug

When using a 2x1x1 supercell of diamond with symmetries, Abinit dectets 24 symmetries instead of 96.
This is because symrel is integer.

The corresponding phonon frequencies computed with anaddb with ifcflag 0 or 1 gives different result.
With 0 no FT is done and the frequencies are correct. Instead with 1, the code give wrong phonon frequencies.

Bug fix
When reconstructing the missing perturbation of the dynamical matrices using symmetries, the code
was using rprim and gprim. These are the same as a primtive cell but symrel now has 2 inside, leading to wrong
symetry matrix in Cartesian coordinate (ss).
The fix is to use rprimd and grimd as done in other part of the code.

Contribution: for this fix I got help from MV,XG,MG.