Bugfix and optimization
A bug was introduced in Abinit v9.8 when updating the format of _EIG2D file to make them real DDB files that can be merged.

In the case of ieig2rf > 2, only the Sternheimer part of the EIG2D should be written to file. This was achieved using the
outbsd routine by copying the eig2nkq array into eig2nkq_tmp in the eig2tot subroutine.
When the outbsd routine was replaced by ddb%set_d2eig_reshape(), the eig2nkq was passed instead of the correct eig2nkq_tmp one,
resulting in a bug --> the fully EIG2D was printed to file instead of the partial Sternheimer only one.
See documentation for explaination https://docs.abinit.org/variables/dfpt/#ieig2rf

This bug is corrected here. In addition, it was not optimal as it could be printed directly in eig2stern, therefore avoiding the dupplication of a large array in eig2nkq_tmp. This has been moved.

I acknowledge useful discussions on this with Xavier Gonze and Gabriel Antonius.


Former-commit-id: d5c9312cf1715e6566624bde29a3536631f56bc5