Enable "use_linalg_gpu" code branch in lobpcg_old with gpu_option==2
Using the new LOBPCG (wfoptalg==114) or CHEBFI (wfoptalg==111)
isn't always possible, because of constraints around paral_kgb,
like Fock operator for example.

For such cases with paral_kgb==0, when a diagonalisation is
needed and desired on GPU, using the old LOBPCG (wfoptalg==14)
is the best option.

Hence, this commit allows the use of old GPU linalg code when using
OpenMP GPU. Thanks to gpu_ wrappers routines, this works on both NVIDIA
and AMD GPUs.