Alloc wfraug after paral_kgb, as it becomes way too big than needed with CHEBFI
MAX(blocksize,ndat) was used here, as prep_fourwf expect
wfraug to be sized after blocksize or ndat depending on paral_kgb.
This becomes lethal when CHEBFi (or LOBPCG with nblock_lobpcg=1) is used
on a big usecase, as blocksize==bandpp*npband.
With this commit, wfraug is always allocated after ndat, which is
sufficient for what we need here.


Former-commit-id: 9ee62852a507c17a333f6a5d06900f5b2ced666b