Fix miscalculation of integer overflow in xmpi_ routines with largetype support
The following is wrong:
```
ntot=int(n1*n2*n3,kind=int64)
```
because the result of `n1*n2*n3` is stored in a standard integer before
being converted to `int64`.
It means an 32bit integer-overflow still happens as the conversion is
perfomed after the multiplication.

This commit fixes with the following:
```
ntot=int(n1,kind=int64)*n2*n3
```
The conversion of `n1` to `int64` also converts the multiplication
to `int64`, meaning its result can have positive values superior to the
maximum for standard 32-bit integers.