Autotools: add 'with-cuda' & 'with-rocm' and deprecate 'with-gpu' & 'with-gpu-flavor'
This change is meant to stick with a more conventionnal approach as to
how user provide root dir of supported GPU libraries.
If user wants to use CUDA, they use '--with-cuda'.
If user wants to use ROCM, they use '--with-rocm'.

Deprecating of 'gpu-flavor' is due to the lack of use of `cuda-simple`.
Flavors `cuda-double` or `hip-double` remain internally used.


Former-commit-id: a230681098eecabe73038c238b86e2155b450320
Former-commit-id: 99d21e759e655d90aa9855d32793c25d66eda1ea