Avoid using a work array in ompgpu_fourwf (except option==3)
`work_gpu` array of ompgpu_fourwf matches size of `fofr`.
However, it appears that those arrays become a huge memory constraint on
huge cases.
On the other hand, we can skip the use of a work array by performing FFT
transforms in-place.
The only offending case is when `option==3` when `fofr` is an input and
isn't expected to be overwritten, so a work array is still needed there.

This commit turns `work_gpu` into a pointer on `fofr` when option/=3,
introduce in-place FFT and arrange some kernels to account for
`fofr`==`work_gpu`, and update GPU memory estimate to account for lesser
memory use in `getghc`.


Former-commit-id: ac99b90b96bca9817d3cfa1d2c9b1c1ce1a5de6f
Former-commit-id: b90e0c68f289cce66af02c6401931db65ad91b14