Update of the Abinit constants to 2022 NIST values
Important
-----------
In 2019, four of the seven SI based unit were defined as exact to avoid to depend on the standard kilogram.
In SI, the Planck constant (h), the elementary electric charge (e), the Boltzmann constant (kB), and the Avogadro constant (NA) are now exact.
I have made a specific subsection to make this clear.
There is a subtlety between the speed of light and fine structure constant. Here I choose to define c.


Former-commit-id: 73392ebd3b8b119a1ee4cabf81eda03373700740