Add a define when GPU unified memory is enabled to disable mapping checks
When a debug flavor is enabled, a check on correction OpenMP mapping is
performed in m_xg.F90... and it breaks in unified memory, which is
harmless in this case.
In order to not remove this check, I added a define to disable OpenMP
mapping checks when unified memory is enabled.


Former-commit-id: acb4186cf1a714630cd429e2fd39175c98440e8a
Former-commit-id: 58fdf6ebf05e653395db15f1bad8ae5b277413d4