Fix performance hit by xomp_get_num_threads (spotted on GH100)
I noticed a massive performance hit in vtowfk when running a DMFT
workload on GH100 with 72 OpenMP threads per task.

After investigating, I realised it was due to a `xomp_get_num_threads`
call performed in `meanvalue_g` inside a double-loop on nband.

The root of the issue is that this routine creates a parallel region to
check the number of threads available, which apparently bears a
stronger overhead on GH100.

This is solved by replacing the call with `xomp_get_max_threads` which
pretty much does the same job, ie giving the max number of threads
available for the next parallel regions.

I made replacements in sections involved with limiting threading (49730d60730)
but it might worth to replace other occurences of `xomp_get_num_threads` as
well...


Former-commit-id: 4b6c606acf0b426d2f94e7902ce4c37abecbdab8